Optimisations
Ive moved some of the computing of the named arguments data to the `SubTag#setNamedArgs` method as they will not change from execution to execution.

These properties are the max & min lengths, as well as the number of important & required arguments after a given position.

Also added som logical optimisations, as well as reducing the amount of code done for some checks to make it more readable